working hibernate mapping
