--- cources.server/softgroup/basis/common-database/common-database-api/src/main/java/com/softgroup/common/database/api/model/Contact.java
@@ -1,16 +1,0 @@
-package com.softgroup.common.database.api.model;
-
-import java.util.UUID;
-
-/**
- * Created by zigza on 16.02.2017.
- */
-public interface Contact {
-    public UUID getId();
-    public String getName();
-    public String getNumber();
-    public String getCity();
-    public void setName(String name);
-    public void setNumber(String number);
-    public void setCity(String city);
-}
