--- cources.server/softgroup/basis/common-database/common-database-api/src/main/java/com/softgroup/common/database/api/model/Conversation.java
@@ -1,19 +1,0 @@
-package com.softgroup.common.database.api.model;
-
-import java.util.Date;
-import java.util.List;
-
-/**
- * Created by zigza on 15.02.2017.
- */
-public interface Conversation {
-    public enum ConversationType {INDIVIDUAL_CHAT,GROUP_CHAT,ALL_CHAT};
-    public String getId();
-    public List<Profile> getParticipants();
-    public Date getDateCreated();
-    public List<Message> getMessages();
-    public ConversationDetails getDetails();
-    public ConversationSettings getSettings();
-    public String getTitle();
-    public ConversationType getType();
-}
