--- cources.server/softgroup/basis/common-database/common-database-api/src/main/java/com/softgroup/common/database/api/model/ConversationDetails.java
@@ -1,8 +1,0 @@
-package com.softgroup.common.database.api.model;
-
-/**
- * Created by zigza on 15.02.2017.
- */
-public interface ConversationDetails {
-    public String getId();
-}
