--- cources.server/softgroup/basis/common-database/common-database-api/src/main/java/com/softgroup/common/database/api/model/ProfileSettings.java
@@ -1,13 +1,0 @@
-package com.softgroup.common.database.api.model;
-
-/**
- * Created by zigza on 15.02.2017.
- */
-public interface ProfileSettings {
-public String getLanguage();
-public String getNickName();
-public Integer getAge();
-public String getAvatar();
-public String getSex();
-public String getId();
-}
