--- cources.server/softgroup/basis/common-database/common-database-impl/pom.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>common-database</artifactId>
-        <groupId>com.softgroup</groupId>
-        <version>0.0.1</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>common-database-impl</artifactId>
-    <dependencies>
-        <dependency>
-            <artifactId>common-database-api</artifactId>
-            <groupId>com.softgroup</groupId>
-        </dependency>
-
-    </dependencies>
-
-</project>
