--- cources.server/softgroup/basis/common-database/common-database-impl/resources/application.properties
@@ -1,13 +1,0 @@
-#Database Configuration
-db.driver=com.mysql.jdbc.Driver
-db.url=jdbc:mysql://127.0.0.1:3306/messenger?useUnicode=yes&characterEncoding=UTF-8&autoReconnect=true&useSSL=false
-db.username=root
-db.password=
-
-#Hibernate Configuration
-hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
-hibernate.hbm2ddl.auto=create-drop
-hibernate.ejb.naming_strategy=org.hibernate.cfg.ImprovedNamingStrategy
-hibernate.show_sql=false
-hibernate.format_sql=true
-hibernate.package_to_scan=com.zigzag.messenger.model
