--- cources.server/softgroup/basis/common-database/common-database-impl/src/main/java/com/softgroup/common/database/impl/model/UserProfileSettings.java
@@ -1,62 +1,0 @@
-package com.softgroup.common.database.impl.model;
-
-import com.softgroup.common.database.api.model.ProfileSettings;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-/**
- * Created by zigza on 21.02.2017.
- */
-@Entity
-public class UserProfileSettings implements ProfileSettings {
-    public UserProfileSettings(){
-
-    }
-
-    @Override
-    public String getLanguage() {
-        return language;
-    }
-
-    @Override
-    public String getNickName() {
-        return nickName;
-    }
-
-    @Override
-    public Integer getAge() {
-        return age;
-    }
-
-    @Override
-    public String getAvatar() {
-        return avatar;
-    }
-
-    @Override
-    public String getSex() {
-        return sex;
-    }
-
-    @Override
-    public String getId() {
-        return id;
-    }
-
-    String language;
-    String nickName;
-    Integer age;
-    String avatar;
-    String sex;
-
-    @Id
-    @GeneratedValue(generator = "hibernate-uuid")
-    @GenericGenerator(name = "uuid", strategy = "uuid2")
-    @Column(name = "id", unique = true)
-    private String id;
-
-}
