+++ cources.server/softgroup/basis/common-database/pom.xml
@@ -10,12 +10,24 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>common-database</artifactId>
-    <packaging>pom</packaging>
-    <modules>
-        <module>common-database-api</module>
-        <module>common-database-impl</module>
-    </modules>
+    <packaging>jar</packaging>
+
     <dependencies>
+        <dependency>
+            <artifactId>common-configuration</artifactId>
+            <groupId>com.softgroup</groupId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-log4j12</artifactId>
+        </dependency>
+
         <dependency>
             <groupId>org.hibernate.javax.persistence</groupId>
             <artifactId>hibernate-jpa-2.0-api</artifactId>
@@ -24,6 +36,11 @@
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-orm</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-core</artifactId>
         </dependency>
 
         <dependency>
@@ -50,6 +67,10 @@
             <groupId>org.hibernate</groupId>
             <artifactId>hibernate-entitymanager</artifactId>
         </dependency>
+
+
+
     </dependencies>
+
 
 </project>
