+++ cources.server/softgroup/basis/common-database/src/main/java/com/softgroup/common/database/model/UserConversationDetails.java
@@ -1,0 +1,27 @@
+package com.softgroup.common.database.model;
+
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+/**
+ * Created by zigza on 20.02.2017.
+ */
+@Entity(name="user_conversation_details")
+public class UserConversationDetails{
+    public UserConversationDetails(){
+
+    }
+    @Id
+    @GeneratedValue(generator = "system-uuid")
+    @GenericGenerator(name = "system-uuid", strategy = "uuid2")
+    @Column(name = "id", unique = true)
+    private String id;
+
+    public String getId() {
+        return id;
+    }
+}
