--- cources.server/softgroup/basis/common-database/src/main/java/com/softgroup/common/database/models/ConversationDetails.java
@@ -1,8 +1,0 @@
-package com.softgroup.common.database.models;
-
-/**
- * Created by zigza on 15.02.2017.
- */
-public interface ConversationDetails {
-    public String getId();
-}
