--- cources.server/softgroup/basis/common-database/src/main/java/com/softgroup/common/database/models/Message.java
@@ -1,16 +1,0 @@
-package com.softgroup.common.database.models;
-
-
-import java.util.Date;
-
-/**
- * Created by zigza on 15.02.2017.
- */
-public interface Message {
-    public String getId();
-    public String getBody();
-    public Profile getAuthor();
-    public Date getDateSended();
-    public Boolean isActive();
-    public Conversation getConversation();
-}
