--- cources.server/softgroup/basis/common-database/src/main/java/com/softgroup/common/database/models/Profile.java
@@ -1,19 +1,0 @@
-package com.softgroup.common.database.models;
-
-import java.util.List;
-
-/**
- * Created by zigza on 15.02.2017.
- */
-public interface Profile {
-    public String getId();
-    public ProfileSettings getSettings();
-    public ProfileStatus getStatus();
-    public List<Message> getMessages();
-    public String getPhoneNumber();
-    public void setPhoneNumber(String phoneNumber);
-    public String getLocaleCode();
-    public void setLocaleCode(String localeCode);
-    public String getDeviceId();
-    public void setDeviceId(String deviceId);
-}
