--- cources.server/softgroup/basis/common-database/src/main/java/com/softgroup/common/database/models/ProfileStatus.java
@@ -1,9 +1,0 @@
-package com.softgroup.common.database.models;
-
-/**
- * Created by zigza on 15.02.2017.
- */
-public interface ProfileStatus {
-    public Boolean isActive();
-    public String getId();
-}
