+++ cources.server/softgroup/basis/pom.xml
@@ -16,6 +16,7 @@
         <module>common-protocol</module>
         <module>common-database</module>
         <module>common-utils</module>
+        <module>common-configuration</module>
 
     </modules>
