+++ cources.server/softgroup/frontend/frontend-controllers/pom.xml
@@ -54,6 +54,11 @@
             <artifactId>authorization-impl</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>com.softgroup</groupId>
+            <artifactId>common-database</artifactId>
+        </dependency>
+
 
     </dependencies>
