--- cources.server/softgroup/frontend/frontend-controllers/src/main/java/com/softgroup/frontend/configurator/PropertiesConfig.java
@@ -1,17 +1,0 @@
-package com.softgroup.frontend.configurator;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.stereotype.Component;
-
-
-/**
- * Created by zigza on 20.02.2017.
- */
-@Component
-public class PropertiesConfig {
-    static @Bean public PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
-    return new PropertySourcesPlaceholderConfigurer();
-    }
-}
-
