+++ cources.server/softgroup/frontend/frontend-controllers/src/main/resources/application.properties
@@ -10,4 +10,4 @@ hibernate.hbm2ddl.auto=create-drop
 hibernate.ejb.naming_strategy=org.hibernate.cfg.ImprovedNamingStrategy
 hibernate.show_sql=false
 hibernate.format_sql=true
-hibernate.package_to_scan=com.zigzag.messenger.model
+hibernate.package_to_scan=com.softgroup.common.database.model
