+++ cources.server/softgroup/pom.xml
@@ -32,6 +32,18 @@
 			<!-- COMMON -->
 
 			<dependency>
+				<groupId>org.slf4j</groupId>
+				<artifactId>slf4j-api</artifactId>
+				<version>1.7.21</version>
+			</dependency>
+
+			<dependency>
+				<groupId>org.slf4j</groupId>
+				<artifactId>slf4j-log4j12</artifactId>
+				<version>1.7.5</version>
+			</dependency>
+
+			<dependency>
 				<groupId>org.springframework</groupId>
 				<artifactId>spring-webmvc</artifactId>
 				<version>${spring.web.version}</version>
@@ -80,8 +92,14 @@
 			</dependency>
 
 			<dependency>
-				<artifactId>common-database-api</artifactId>
 				<groupId>com.softgroup</groupId>
+				<artifactId>common-database</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+
+			<dependency>
+				<groupId>com.softgroup</groupId>
+				<artifactId>common-configuration</artifactId>
 				<version>${project.version}</version>
 			</dependency>
